Move tests to pattern structure
