+++ cloud/wfm/test/mocha.opts
@@ -1,3 +1,3 @@
 --compilers ts:ts-node/register
 --require source-map-support/register
-test/**.ts
+test/**/*.ts
