+++ cloud/wfm/test/process/ProcessImpl.ts
@@ -1,0 +1,6 @@
+import ProcessImpl from '../../src/process';
+import suite from './index';
+
+describe('ProcessImpl', function() {
+  suite(name => new ProcessImpl(name));
+});
